Finish refactoring convert_to except obj
