+++ crates/sui-graphql-rpc/src/types/checkpoint.rs
@@ -1,8 +1,12 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use super::{base64::Base64, end_of_epoch_data::EndOfEpochData, epoch::Epoch, gas::GasCostSummary};
+use super::{
+    base64::Base64, committee_member::CommitteeMember, end_of_epoch_data::EndOfEpochData,
+    epoch::Epoch, gas::GasCostSummary,
+};
 use async_graphql::*;
+use fastcrypto::traits::EncodeDecodeBase64;
 
 #[derive(Clone, Debug, PartialEq, Eq, SimpleObject)]
 pub(crate) struct Checkpoint {
@@ -15,8 +19,52 @@ pub(crate) struct Checkpoint {
     pub live_object_set_digest: Option<String>,
     pub network_total_transactions: Option<u64>,
     pub rolling_gas_summary: Option<GasCostSummary>,
-    pub epoch: Option<Epoch>,
+    pub epoch: Epoch,
     pub end_of_epoch: Option<EndOfEpochData>,
     // transactionConnection(first: Int, after: String, last: Int, before: String): TransactionBlockConnection
     // address_metrics: AddressMetrics,
+}
+
+impl From<&sui_json_rpc_types::Checkpoint> for Checkpoint {
+    fn from(c: &sui_json_rpc_types::Checkpoint) -> Self {
+        let end_of_epoch_data = &c.end_of_epoch_data;
+        let end_of_epoch = end_of_epoch_data.clone().map(|e| {
+            let committees = e.next_epoch_committee;
+            let new_committee = if committees.is_empty() {
+                None
+            } else {
+                Some(
+                    committees
+                        .iter()
+                        .map(|c| CommitteeMember {
+                            authority_name: Some(c.0.into_concise().to_string()),
+                            stake_unit: Some(c.1),
+                        })
+                        .collect::<Vec<_>>(),
+                )
+            };
+
+            EndOfEpochData {
+                new_committee,
+                next_protocol_version: Some(e.next_epoch_protocol_version.as_u64()),
+            }
+        });
+
+        Self {
+            digest: c.digest.to_string(),
+            sequence_number: c.sequence_number,
+            validator_signature: Some(Base64::from(
+                c.validator_signature.encode_base64().into_bytes(),
+            )),
+            previous_checkpoint_digest: c.previous_digest.map(|x| x.to_string()),
+            live_object_set_digest: None, // TODO fix this
+            network_total_transactions: Some(c.network_total_transactions),
+            rolling_gas_summary: Some(GasCostSummary::from(&c.epoch_rolling_gas_cost_summary)),
+            epoch: Epoch {
+                epoch_id: c.epoch,
+                gas_cost_summary: Some(GasCostSummary::from(&c.epoch_rolling_gas_cost_summary)),
+            },
+            end_of_epoch,
+        }
+    }
 }
